move ArrayPushCallbackChunkFormat into separate plugin
