+++ lib/web/JsonpHelpers.js
@@ -10,6 +10,8 @@
 /** @typedef {import("../ChunkGroup")} ChunkGroup */
 /** @typedef {(string|number)[]} EntryItem */
 
+// TODO move to this file to ../javascript/ChunkHelpers.js
+
 /**
  * @param {ChunkGroup} chunkGroup a chunk group
  * @returns {Set<Chunk>} chunks
