+++ lib/web/JsonpTemplatePlugin.js
@@ -6,15 +6,12 @@
 "use strict";
 
 const { SyncWaterfallHook } = require("tapable");
-const { ConcatSource } = require("webpack-sources");
 const Compilation = require("../Compilation");
-const HotUpdateChunk = require("../HotUpdateChunk");
 const RuntimeGlobals = require("../RuntimeGlobals");
 const Template = require("../Template");
-const JavascriptModulesPlugin = require("../javascript/JavascriptModulesPlugin");
-const chunkHasJs = require("../javascript/JavascriptModulesPlugin").chunkHasJs;
+const ArrayPushCallbackChunkFormatPlugin = require("../javascript/ArrayPushCallbackChunkFormatPlugin");
 const JsonpChunkLoadingRuntimeModule = require("./JsonpChunkLoadingRuntimeModule");
-const { getEntryInfo, needEntryDeferringCode } = require("./JsonpHelpers");
+const { needEntryDeferringCode } = require("./JsonpHelpers");
 
 /** @typedef {import("../Chunk")} Chunk */
 /** @typedef {import("../Compilation")} Compilation */
@@ -57,71 +54,8 @@ class JsonpTemplatePlugin {
 	 * @returns {void}
 	 */
 	apply(compiler) {
+		new ArrayPushCallbackChunkFormatPlugin().apply(compiler);
 		compiler.hooks.thisCompilation.tap("JsonpTemplatePlugin", compilation => {
-			const hooks = JavascriptModulesPlugin.getCompilationHooks(compilation);
-			hooks.renderChunk.tap("JsonpTemplatePlugin", (modules, renderContext) => {
-				const { chunk, chunkGraph, runtimeTemplate } = renderContext;
-				const hotUpdateChunk = chunk instanceof HotUpdateChunk ? chunk : null;
-				const globalObject = runtimeTemplate.outputOptions.globalObject;
-				const source = new ConcatSource();
-				const runtimeModules = chunkGraph.getChunkRuntimeModulesInOrder(chunk);
-				const runtimePart =
-					runtimeModules.length > 0 &&
-					Template.renderChunkRuntimeModules(runtimeModules, renderContext);
-				if (hotUpdateChunk) {
-					const jsonpFunction = runtimeTemplate.outputOptions.hotUpdateFunction;
-					source.add(`${globalObject}[${JSON.stringify(jsonpFunction)}](`);
-					source.add(`${JSON.stringify(chunk.id)},`);
-					source.add(modules);
-					if (runtimePart) {
-						source.add(",\n");
-						source.add(runtimePart);
-					}
-					source.add(")");
-				} else {
-					const jsonpFunction = runtimeTemplate.outputOptions.jsonpFunction;
-					source.add(
-						`(${globalObject}[${JSON.stringify(
-							jsonpFunction
-						)}] = ${globalObject}[${JSON.stringify(
-							jsonpFunction
-						)}] || []).push([`
-					);
-					source.add(`${JSON.stringify(chunk.ids)},`);
-					source.add(modules);
-					const entries = getEntryInfo(chunkGraph, chunk, c =>
-						chunkHasJs(c, chunkGraph)
-					);
-					const entriesPart =
-						entries.length > 0 && `,${JSON.stringify(entries)}`;
-					if (entriesPart || runtimePart) {
-						source.add(entriesPart || ",0");
-					}
-					if (runtimePart) {
-						source.add(",\n");
-						source.add(runtimePart);
-					}
-					source.add("])");
-				}
-				return source;
-			});
-			hooks.chunkHash.tap(
-				"JsonpTemplatePlugin",
-				(chunk, hash, { chunkGraph, runtimeTemplate }) => {
-					if (chunk.hasRuntime()) return;
-					hash.update("JsonpTemplatePlugin");
-					hash.update("1");
-					hash.update(
-						JSON.stringify(
-							getEntryInfo(chunkGraph, chunk, c => chunkHasJs(c, chunkGraph))
-						)
-					);
-					hash.update(`${runtimeTemplate.outputOptions.jsonpFunction}`);
-					hash.update(`${runtimeTemplate.outputOptions.hotUpdateFunction}`);
-					hash.update(`${runtimeTemplate.outputOptions.globalObject}`);
-				}
-			);
-
 			const {
 				linkPreload,
 				linkPrefetch
